+++ examples/common/metrics_foo_library/foo_library.h
@@ -12,6 +12,7 @@ class foo_library
 public:
   static void counter_example(const std::string &name);
   static void histogram_example(const std::string &name);
+  static void histogram_exp_example(const std::string &name);
   static void observable_counter_example(const std::string &name);
 #if OPENTELEMETRY_ABI_VERSION_NO >= 2
   static void gauge_example(const std::string &name);