+++ sdk/include/opentelemetry/sdk/metrics/aggregation/aggregation.h
@@ -3,6 +3,7 @@
 
 #pragma once
 
+#include <stdint.h>
 #include <memory>
 
 #include "opentelemetry/version.h"