+++ CHANGELOG.md
@@ -15,6 +15,15 @@ Increment the:
 
 ## [Unreleased]
 
+* [SDK] Base2 exponential histogram aggregation
+  [#3175](https://github.com/open-telemetry/opentelemetry-cpp/pull/3346)
+
+  New Features:
+
+  Add base2 exponential histogram aggregation. Includes a new aggregation type,
+  ostream exporter, and otlp/grpc exporter. Updated histogram aggregation and
+  benchmark tests.
+
 * [API] Remove `WITH_ABSEIL` and `HAVE_ABSEIL`
   [#3318](https://github.com/open-telemetry/opentelemetry-cpp/pull/3318)